Module rsim::processor::elements::memory

Structs

AggregateMemory
Struct that combines a set of array-backed memory mappings. The mapped address ranges may not overlap.

IOMemory
I/O Memory Defines an address range of a single u32. Reads from this address return 0, Writes to this address throw ProgramHaltedException::ResultReturned.

MemoryBacking
Array-backed memory

Traits

Memory
Public trait which supplies {load,store}_u{8,16,32,64} functions.

MemoryOf
Internal trait defining functions for reading/writing values of type TData from a memory, using values of type TAddr as addresses (TAddr is usually u64 for non-CHERI memory.)

Functions

check_alignment_range
Checks for address alignment and whether the address is in-range, generic on the type of data being accessed. Enforces alignment to size_of.

Type Definitions

MemoryResult