Trait rsim::processor::isa_mods::IsaMod

pub trait IsaMod<TConn> {
    type Pc;
    fn will_handle(&self, opcode: Opcode, inst: InstructionBits) -> bool;

    fn execute(
        &mut self, 
        opcode: Opcode, 
        inst: InstructionBits, 
        inst_bits: u32, 
        conn: TConn
    ) -> ProcessorResult<Option<Self::Pc>>;
}

Trait for all ISA modules.

Uses a TConn value to connect to processor state: for example, the Rv64im module takes a connection holding a reference to a 64-bit register file, and a reference to main memory.

Associated Types

type Pc

Loading content...
Required methods

fn will_handle(&self, opcode: Opcode, inst: InstructionBits) -> bool

Return true if this ISA module should handle the given instruction

fn execute(
&mut self,
opcode: Opcode,
inst: InstructionBits,
inst_bits: u32,
conn: TConn
) -> ProcessorResult<Option<Self::Pc>>

Execute the given instruction, returning the new PC (or None if continuing as normal)

Loading content...