Trait rsim::processor::isa_mods::vector::VecMemInterface

pub trait VecMemInterface<uXLEN: PossibleXlen, TElem> {
    fn load_from_memory(
        &mut self, 
        eew: Sew, 
        addr_provenance: (u64, Provenance)
    ) -> Result<TElem>;

    fn store_to_memory(
        &mut self, 
        eew: Sew, 
        val: TElem, 
        addr_provenance: (u64, Provenance)
    ) -> Result<()>;
}

Common trait for interfaces to memory

Required methods

fn load_from_memory(
&mut self,
eew: Sew,
addr_provenance: (u64, Provenance)
) -> Result<TElem>

Use an address, provenance pair to read a vector element from memory

fn store_to_memory(
&mut self,
eew: Sew,
val: TElem,
addr_provenance: (u64, Provenance)
) -> Result<()>

Use an address, provenance pair to write a vector element to memory

Loading content...