Module rsim::simulation

A collection of restaurant simulation representations

Structs

Configuration
A configuration for a restaurant simuation

Driver
A driver going to pick up an order

Order
An order of food

RestaurantSimulation
A simulation of a restaurant where orders are created and drivers pickup and deliver

Shelf
A shelf that holds a limited amount of items

ShelfItem
An item on the shelf

Constants

COLD
String representation of cold temperature

FROZEN
String representation of frozen temperature

HOT
String representation of hot temperature

SHELF_COLD
Shelf index for cold items

SHELF_FROZEN
Shelf index for frozen items

SHELF_HOT
Shelf index for hot items

SHELF_OVERFLOW
Shelf index for overflow items