Module rsim::wasm

The glue between javascript and web assembly to run a restaurant simulation

Structs

SIMULATION
Global restaurant simulation instance.

Functions

error
Log an error message

log
Log a str to the screen

malloc
Allocate some memory in our web assembly module from javascript

random_f32
Generate a random f32

rsim_error⚠
Log a an error C cstring to javascript

rsim_log⚠
Log a C cstring to javascript

rsim_rand⚠
Generate a random f32 from javascript

rsim_set_view_items⚠
Send the shelf items as json to javascript

set_view_items
Log an error message